+++ packages/emissions-dashboard/package.json
@@ -50,7 +50,7 @@
     "styled-components": "^5.2.1",
     "svg-loaders-react": "^2.2.1",
     "tailwindcss": "^2.0.4",
-    "time-series-aggregator": "^1.0.1",
+    "time-series-aggregator": "^1.1.0",
     "typechain": "^5.0.0",
     "web3": "^1.3.1",
     "web3-react": "^5.0.5"
